% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{boot_dcm}
\alias{boot_dcm}
\title{Confidence interval of the reliability (accuracy index)}
\usage{
boot_dcm(model, v_dep, v_ind = NULL, v_ind_composites = NULL,
  sample_size = 1000, replications = 1000)
}
\arguments{
\item{model}{Population relations among variables represented by Lavaan Syntax}

\item{v_dep}{Vector of names of the dependent variables in your profile.}

\item{v_ind}{Vector of names of independent variables you would like to control for.}

\item{v_ind_composites}{Vector of names of independent variables that are composites of dependent variables}

\item{sample_size}{The number of cases}

\item{replications}{The number of trials}
}
\value{
simulated 95% confidence interval
}
\description{
Confidence interval of the reliability (accuracy index)
}
\examples{
m <- "
Gc =~ 0.85 * Gc_1 + 0.68 * Gc_2 + 0.80 * Gc_3
Gf =~ 0.80 * Gf_1 + 0.90 * Gf_2 + 0.80 * Gf_3
Gs =~ 0.70 * Gs_1 + 0.80 * Gs_2 + 0.80 * Gs_3
Read =~ 0.66 * Read_1 + 0.85 * Read_2 + 0.91 * Read_3
Math =~ 0.40 * Math_1 + 0.90 * Math_2 + 0.70 * Math3
Gc ~ 0.60 * Gf + 0.10 * Gs
Gf ~ 0.50 * Gs
Read ~ 0.40 * Gc + 0.10 * Gf
Math ~ 0.20 * Gc + 0.30 * Gf + 0.10 * Gs
"
boot_dcm(model = m,
         v_dep = c("Math", "Read"),
         v_ind = c("Gc", "Gf", "Gs"),
         sample_size = 100,
         replications = 100)
}
