% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{CM_cor}
\alias{CM_cor}
\title{Automatically calculate the accuracy of one condition.}
\usage{
CM_cor(Dep, Ind = NULL, d)
}
\arguments{
\item{Dep}{The names of variables you would like to condition on.}

\item{Ind}{The names of variables of your interest.}

\item{d}{Dataset created by using simStandarized.}
}
\value{
Correlation between the conditional Mahalanobis distance calculated by using the true scores and the conditional Mahalanobis calculated by using estimated factor scores
}
\description{
Automatically calculate the accuracy of one condition.
}
\examples{
SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
d_demo <- simStandardized(SimModel, 10)
CM_cor(Dep = c("Math", "Read"),Ind = c("Gc", "Gf", "Gs") , d_demo)
}
