% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CMahalanobis}
\alias{CMahalanobis}
\title{Calculate the conditional Mahalanobis distance for any variables.}
\usage{
CMahalanobis(Dep, Ind = NULL, R, d)
}
\arguments{
\item{Dep}{The names of variables you would like to condition on.}

\item{Ind}{The names of variables of your interest.}

\item{R}{Conditional correlation among variables.}

\item{d}{Profiles of interest.}
}
\value{
conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.
}
\description{
Calculate the conditional Mahalanobis distance for any variables.
}
\examples{
SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
d_demo <- simStandardized(SimModel, 10)
CMahalanobis(c("Math", "Read"),c("Gf", "Gs", "Gc"),d_demo$R_all,d_demo$Data)
}
