% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_example}
\alias{d_example}
\title{An example data.frame}
\format{
A data frame with 1 row and 8 variables:
\describe{
\item{X_1}{A predictor variable}
\item{X_2}{A predictor variable}
\item{X_3}{A predictor variable}
\item{Y_1}{An outcome variable}
\item{Y_2}{An outcome variable}
\item{Y_3}{An outcome variable}
\item{X}{A latent predictor variable}
\item{Y}{A latent outcome variable}
}
}
\usage{
d_example
}
\description{
A dataset with 1 row of data for a single case.
}
\keyword{datasets}
