% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dCM_cor}
\alias{dCM_cor}
\title{Function to evaluate the accuracy of dCM with estimated factor scores}
\usage{
dCM_cor(m, Dep, Ind, n = 10000)
}
\arguments{
\item{m}{Lavaan Syntax Object}

\item{Dep}{The names of variables you would like to condition on.}

\item{Ind}{The names of variables of your interest.}

\item{n}{Sample size of simulated data}
}
\value{
Correlation between the conditional Mahalanobis distance calculated by using the true scores and the conditional Mahalanobis calculated by using estimated factor scores
}
\description{
Function to evaluate the accuracy of dCM with estimated factor scores
}
\examples{
m <- "
Gc =~ 0.85 * Gc_1 + 0.68 * Gc_2 + 0.80 * Gc_3
Gf =~ 0.80 * Gf_1 + 0.90 * Gf_2 + 0.80 * Gf_3
Gs =~ 0.70 * Gs_1 + 0.80 * Gs_2 + 0.80 * Gs_3
Read =~ 0.66 * Read_1 + 0.85 * Read_2 + 0.91 * Read_3
Math =~ 0.40 * Math_1 + 0.90 * Math_2 + 0.70 * Math_3
Gc ~ 0.60 * Gf + 0.10 * Gs
Gf ~ 0.50 * Gs
Read ~ 0.40 * Gc + 0.10 * Gf
Math ~ 0.20 * Gc + 0.30 * Gf + 0.10 * Gs
"
dCM_cor(m, Dep = c("Math", "Read"), Ind = c("Gc", "Gf", "Gs"), n = 100)
}
