% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estStandardized}
\alias{estStandardized}
\title{Estimate factor scores for a given profile and population model.}
\usage{
estStandardized(m, d)
}
\arguments{
\item{m}{Lavaan Syntax Object.}

\item{d}{Observed Z scores.}
}
\value{
Standardized estimated factor scores.
}
\description{
Estimate factor scores for a given profile and population model.
}
\examples{
m = "LatantVariable =~ 0.8 * ObservedVar1 + 0.8 * ObservedVar2 + 0.8 * Observed3"
d <- simStandardized(m, 1000) #simulate 1000 cases (profiles)

# Pick up the first three cases for illustration
demo_individual_exp <- d$Data[1:3,]

# Estimate factor scores based on the three cases as well as the population model in m
estStandardized(m,  demo_individual_exp)
}
