% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{glue2simf}
\alias{glue2simf}
\title{Function to stuck parameters to the lavaan model}
\usage{
glue2simf(SimModel, para)
}
\arguments{
\item{SimModel}{Population relations among variables represented by Lavaan Syntax.}

\item{para}{Parameters for simulation studies.}
}
\value{
A list of lavaan object with specified parameters.
}
\description{
Function to stuck parameters to the lavaan model
}
