% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mahaName}
\alias{mahaName}
\title{Extract names from a lavaan syntax object.}
\usage{
mahaName(m)
}
\arguments{
\item{m}{Lavaan Syntax Object.}
}
\value{
Names of observed variables, latent variables.
}
\description{
Extract names from a lavaan syntax object.
}
\examples{
m = "LatantVariable =~ ObservedVar1 + ObservedVar2 + Observed3"
mahaName(m)
}
