% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{makePara}
\alias{makePara}
\title{function to make a set of parameters that follows beta distribution}
\usage{
makePara(from, to, n, k)
}
\arguments{
\item{from}{starting point for the parameters}

\item{to}{ending point for the parameters}

\item{n}{number of cases}

\item{k}{number of variables.}
}
\value{
all parameters, lavaan object with specified parameters and data generated for each condition by calling simStandarized and the accuracy index for each condition
}
\description{
function to make a set of parameters that follows beta distribution
}
\examples{
makePara(1000, 9000, n = 10, k = 15)
}
