% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{p2label}
\alias{p2label}
\title{Range label associated with probability}
\usage{
p2label(p)
}
\arguments{
\item{p}{Probability}
}
\value{
label string
}
\description{
Range label associated with probability
}
\keyword{internal}
