% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.cond_maha}
\alias{plot.cond_maha}
\title{Plot the variables from the results of the cond_maha function.}
\usage{
\method{plot}{cond_maha}(
  cm,
  p_tail = 0,
  family = "serif",
  score_digits = ifelse(min(cm$sigma) >= 10, 0, 2)
)
}
\arguments{
\item{cm}{The results of the cond_maha function.}

\item{family}{Font family.}

\item{score_digits}{Number of digits to round scores.}
}
\description{
Plot the variables from the results of the cond_maha function.
}
