% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.maha}
\alias{plot.maha}
\title{Plot objects of the maha class (i.e, the results of the cond_maha function
using dependent variables only).}
\usage{
\method{plot}{maha}(
  cm,
  p_tail = 0,
  family = "serif",
  score_digits = ifelse(min(cm$sigma) >= 10, 0, 2)
)
}
\arguments{
\item{cm}{The results of the cond_maha function.}

\item{p_tail}{Proportion in violin tail (defaults to 0).}

\item{family}{Font family.}

\item{score_digits}{Number of digits to round scores.}
}
\description{
Plot objects of the maha class (i.e, the results of the cond_maha function
using dependent variables only).
}
