% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Sim2Free}
\alias{Sim2Free}
\title{Function that removes all fixed values in a lavaan model}
\usage{
Sim2Free(m)
}
\arguments{
\item{m}{Population relations among variables represented by Lavaan Syntax.}
}
\value{
Lavaan object without specified coefficients
}
\description{
Function that removes all fixed values in a lavaan model
}
\examples{
SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
Sim2Free(SimModel)
}
