% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simCor}
\alias{simCor}
\title{Simulate data and cor between true conditional Mahalanobis distance and the estimated}
\usage{
simCor(dPar, SimModel, Dep, Ind = NULL, n = 100)
}
\arguments{
\item{dPar}{Parameters we would like to test in simulation studies}

\item{SimModel}{Population relations among variables represented by Lavaan Syntax}

\item{Dep}{The names of variables you would like to condition on}

\item{Ind}{The names of variables of your interest}

\item{n}{The number of cases}
}
\value{
all parameters, lavaan object with specified parameters and data generated for each condition by calling simStandarized and the accuracy index for each condition
}
\description{
Simulate data and cor between true conditional Mahalanobis distance and the estimated
}
\examples{
SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
beta_measure1 <- makePara(1000, 9000, n = 10, k = 15)
beta_structure1 <- makePara(1000, 1000, n = 10, k = 8)
par_beta1 <- cbind(beta_measure1, beta_structure1)
colnames(par_beta1) <- paste0("x", 1:23)
simCor(dPar = par_beta1, SimModel, Dep = c("Read", "Math"),Ind = c("Gc", "Gs", "Gf"), n = 10)
}
