% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simMaha}
\alias{simMaha}
\title{Create a tibble to hold everything for simulation}
\usage{
simMaha(dPar, SimModel, n = 100)
}
\arguments{
\item{dPar}{Parameters we would like to test in simulation studies}

\item{SimModel}{Population relations among variables represented by Lavaan Syntax}

\item{n}{The number of cases}
}
\value{
all parameters, lavaan object with specified parameters and data genereged for each condition by calling simStandarized
}
\description{
Create a tibble to hold everything for simulation
}
\examples{
SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
beta_measure1 <- makePara(1000, 9000, n = 10, k = 15)
beta_structure1 <- makePara(1000, 1000, n = 10, k = 8)
cbind(beta_measure1, beta_structure1)-> par_beta1
colnames(par_beta1) <- paste0("x", 1:23)
simMaha(par_beta1, SimModel, n = 10)
}
