% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{unusualness}
\alias{unusualness}
\title{Wrapper for finding out Mahalanobis distance between variables: this one gives everything for practitioners to use when they only have population relations and their clients' data}
\usage{
unusualness(data, model, v_dep, v_ind)
}
\arguments{
\item{data}{Profiles of interest.}

\item{model}{Structural model represented by lavaan Syntax.}

\item{v_dep}{The names of variables you would like to condition on.}

\item{v_ind}{The names of variables of your interest.}
}
\value{
conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.
}
\description{
Wrapper for finding out Mahalanobis distance between variables: this one gives everything for practitioners to use when they only have population relations and their clients' data
}
\examples{
# Standardized structural model in lavaan syntax
m <- "
Gc =~ 0.85 * Gc_1 + 0.68 * Gc_2 + 0.80 * Gc_3
Gf =~ 0.80 * Gf_1 + 0.90 * Gf_2 + 0.80 * Gf_3
Read =~ 0.66 * Read_1 + 0.85 * Read_2 + 0.91 * Read_3
Math =~ 0.40 * Math_1 + 0.90 * Math_2 + 0.70 * Math_3
Gc ~ 0.60 * Gf
Read ~ 0.40 * Gc + 0.10 * Gf
Math ~ 0.20 * Gc + 0.30 * Gf
"

# Put observed scores in data.frame
d_demo <- data.frame(
          Gc_1 = -1,
          Gc_2 = 0.5,
          Gc_3 = -0.2,
          Gf_1 = 1.1,
          Gf_2 = 1.3,
          Gf_3 = 2,
          Read_1 = -0.5,
          Read_2 = -1,
          Read_3 = -1.4,
          Math_1 = 1.1,
          Math_2 = 1.3,
          Math_3 = 0.7
)
unusualness(
     data = d_demo,
     model = m,
     v_dep = c("Math", "Read"),
     v_ind = c("Gc", "Gf"))
}
