% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unusualprofile-package.R
\docType{package}
\name{unusualprofile-package}
\alias{unusualprofile}
\alias{unusualprofile-package}
\title{unusualprofile: Calculates Conditional Mahalanobis Distances}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculates a Mahalanobis distance for every row of a set of outcome variables (Mahalanobis, 1936 \doi{10.1007/s13171-019-00164-5}). The conditional Mahalanobis distance is calculated using a conditional covariance matrix (i.e., a covariance matrix of the outcome variables after controlling for a set of predictors). Plotting the output of the cond_maha() function can help identify which elements of a profile are unusual after controlling for the predictors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/unusualprofile}
  \item Report bugs at \url{https://github.com/wjschne/unusualprofile/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

Authors:
\itemize{
  \item Feng Ji \email{fengji@berkeley.edu}
}

}
\keyword{internal}
