% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unusualprofile-package.R
\docType{package}
\name{unusualprofile}
\alias{unusualprofile}
\title{unusualprofile: Calculates Conditional Mahalanobis Distances}
\description{
The unusualprofile package calculates the unusualness of score profiles
conditioned on a set of predictor variables
}
\author{
W. Joel Schneider

Feng Ji
}
