% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estStandardized}
\alias{estStandardized}
\title{Estimate factor scores for a given profile and population model.}
\usage{
estStandardized(d, m)
}
\arguments{
\item{d}{observed z-scores in matrix or data.frame}

\item{m}{Structural model represented by lavaan Syntax}
}
\value{
simStandardized output list with estimated factor scores from d
}
\description{
Estimate factor scores for a given profile and population model.
}
\examples{
m = "latent_1 =~ 0.8 * ob_1 + 0.8 * ob_2 + 0.8 * ob_3"
d <- data.frame(ob_1 = 1, ob_2 = -0.5, ob_3 = 1.2)

# Estimate factor scores based on this case
estStandardized(d = d, m = m)
}
