% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.cond_maha}
\alias{plot.cond_maha}
\title{Plot the variables from the results of the cond_maha function.}
\usage{
\method{plot}{cond_maha}(
  x,
  ...,
  p_tail = 0,
  family = "sans",
  score_digits = ifelse(min(x$sigma) >= 10, 0, 2)
)
}
\arguments{
\item{x}{The results of the cond_maha function.}

\item{...}{Arguments passed to print function}

\item{p_tail}{The proportion of the tail to shade}

\item{family}{Font family.}

\item{score_digits}{Number of digits to round scores.}
}
\value{
A ggplot2-object
}
\description{
Plot the variables from the results of the cond_maha function.
}
