% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simStandardized}
\alias{simStandardized}
\title{Function that takes a lavaan model with standardized parameters and simulates latent scores, errors, disturbances, and observed scores}
\usage{
simStandardized(m, n = 1000, ObservedOnly = FALSE)
}
\arguments{
\item{m}{Structural model represented by lavaan Syntax}

\item{n}{Number of simulated cases}

\item{ObservedOnly}{Return only observed data}
}
\value{
Latent scores, errors, disturbances, and observed scores
}
\description{
Function that takes a lavaan model with standardized parameters and simulates latent scores, errors, disturbances, and observed scores
}
\examples{
# Lavaan model
m = "Latent_1 =~ 0.8 * Ob_1 + 0.8 * Ob_2"

# simulate 10 cases
simStandardized(m, n = 10)
}
