% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unusualprofile.R
\docType{package}
\name{unusualprofile}
\alias{unusualprofile}
\alias{unusualprofile-package}
\title{unusualprofile: Calculates Conditional Mahalanobis Distances}
\description{
The unusualprofile package calculates the unusualness of score profiles conditioned on
a set of predictor variables
}
\section{Foo functions}{

The foo functions ...
}

\author{
Feng Ji

W. Joel Schneider
}
